Moving tqec.scale to tqec.utils.scale
