+++ src/tqec/circuit/qubit_map.py
@@ -19,7 +19,7 @@ import stim
 
 from tqec.circuit.qubit import GridQubit
 from tqec.exceptions import TQECException
-from tqec.scale import round_or_fail
+from tqec.utils.scale import round_or_fail
 
 
 @dataclass(frozen=True)
