+++ src/tqec/compile/block.py
@@ -9,10 +9,10 @@ from tqec.exceptions import TQECException
 from tqec.plaquette.library.empty import empty_square_plaquette
 from tqec.plaquette.plaquette import Plaquette, Plaquettes, RepeatedPlaquettes
 from tqec.position import BlockPosition2D, Shift2D
-from tqec.scale import LinearFunction
 from tqec.templates.indices.base import RectangularTemplate
 from tqec.templates.indices.layout import LayoutTemplate
 from tqec.utils.frozendefaultdict import FrozenDefaultDict
+from tqec.utils.scale import LinearFunction
 
 
 @dataclass
