+++ src/tqec/compile/compile.py
@@ -34,9 +34,9 @@ from tqec.exceptions import TQECException, TQECWarning
 from tqec.noise_model import NoiseModel
 from tqec.plaquette.plaquette import Plaquettes, RepeatedPlaquettes
 from tqec.position import Direction3D, Position3D
-from tqec.scale import round_or_fail
 from tqec.templates.indices.base import Template
 from tqec.templates.indices.layout import LayoutTemplate
+from tqec.utils.scale import round_or_fail
 
 
 @dataclass
