+++ src/tqec/compile/specs/library/_utils.py
@@ -8,9 +8,9 @@ from tqec.plaquette.enums import PlaquetteOrientation, PlaquetteSide
 from tqec.plaquette.library import PlaquetteBuilder, empty_square_plaquette
 from tqec.plaquette.plaquette import Plaquette, Plaquettes
 from tqec.position import Direction3D
-from tqec.scale import LinearFunction
 from tqec.templates.indices.qubit import QubitTemplate
 from tqec.utils.frozendefaultdict import FrozenDefaultDict
+from tqec.utils.scale import LinearFunction
 
 _DEFAULT_BLOCK_REPETITIONS = LinearFunction(2, -1)
