+++ src/tqec/computation/zx_graph.py
@@ -14,7 +14,7 @@ from mpl_toolkits.mplot3d.axes3d import Axes3D
 from tqec.computation._base_graph import ComputationGraph
 from tqec.exceptions import TQECException
 from tqec.position import Direction3D, Position3D
-from tqec.scale import round_or_fail
+from tqec.utils.scale import round_or_fail
 
 if TYPE_CHECKING:
     from tqec.computation.block_graph import BlockGraph
