+++ src/tqec/plaquette/plaquette.py
@@ -12,8 +12,8 @@ from tqec.exceptions import TQECException
 from tqec.plaquette.enums import PlaquetteOrientation
 from tqec.plaquette.qubit import PlaquetteQubits
 from tqec.position import PhysicalQubitPosition2D
-from tqec.scale import LinearFunction, round_or_fail
 from tqec.utils.frozendefaultdict import FrozenDefaultDict
+from tqec.utils.scale import LinearFunction, round_or_fail
 
 
 @dataclass(frozen=True)
