+++ src/tqec/templates/indices/_testing.py
@@ -5,9 +5,9 @@ import numpy.typing as npt
 from typing_extensions import override
 
 from tqec.position import Shift2D
-from tqec.scale import LinearFunction, Scalable2D
 from tqec.templates.indices.base import BorderIndices, RectangularTemplate
 from tqec.templates.indices.enums import TemplateBorder
+from tqec.utils.scale import LinearFunction, Scalable2D
 
 
 class FixedTemplate(RectangularTemplate):
