+++ src/tqec/templates/indices/layout.py
@@ -73,8 +73,8 @@ from typing_extensions import override
 
 from tqec.exceptions import TQECException
 from tqec.position import BlockPosition2D, PlaquettePosition2D, Shape2D, Shift2D
-from tqec.scale import Scalable2D
 from tqec.templates.indices.base import RectangularTemplate, Template
+from tqec.utils.scale import Scalable2D
 
 
 class LayoutTemplate(Template):
