+++ src/tqec/templates/indices/qubit.py
@@ -8,9 +8,9 @@ import numpy.typing as npt
 from typing_extensions import override
 
 from tqec.exceptions import TQECException, TQECWarning
-from tqec.scale import LinearFunction, Scalable2D
 from tqec.templates.indices.base import BorderIndices, RectangularTemplate
 from tqec.templates.indices.enums import TemplateBorder
+from tqec.utils.scale import LinearFunction, Scalable2D
 
 
 class QubitTemplate(RectangularTemplate):
