+++ src/tqec/utils/__init__.py
@@ -1,0 +1,3 @@
+from .scale import LinearFunction as LinearFunction
+from .scale import Scalable2D as Scalable2D
+from .scale import round_or_fail as round_or_fail
