+++ src/tqec/utils/scale_test.py
@@ -2,7 +2,7 @@ import pytest
 
 from tqec.exceptions import TQECException
 from tqec.position import Shape2D
-from tqec.scale import LinearFunction, Scalable2D, round_or_fail
+from tqec.utils.scale import LinearFunction, Scalable2D, round_or_fail
 
 
 @pytest.mark.parametrize(
